check for column before attempting to remove

+++ db/migrate/20160806024245_drop_campaign_id_from_rides.rb
@@ -1,5 +1,7 @@
 class DropCampaignIdFromRides < ActiveRecord::Migration[5.0]
   def change
-    remove_column :rides, :campaign_id, :integer
+    if column_exists?(:rides, :campaign_id)
+      remove_column :rides, :campaign_id, :integer
+    end
   end
 end
